+++ include/interactable.h
@@ -30,38 +30,38 @@ inline bool objCheckSavePart(const std::string part, const ObjType objType)
 class Interactable
 {
 public:
-    Interactable(const Vect<int> tileSize, const std::vector<uint8_t> color, const ObjType type);
+    Interactable(const Vect<uint32_t> tileSize, const std::vector<uint8_t> color, const ObjType type);
     Interactable(const ObjType type); // For loading from save
     virtual ~Interactable();
 
     void operator=(const Interactable&) = delete;
 
-    virtual bool canPlace(const Vect<int>& pos, std::vector<std::unique_ptr<Interactable>>& objects, const Vect<int>& size);
+    virtual bool canPlace(const Vect<int64_t>& pos, std::vector<std::unique_ptr<Interactable>>& objects, const Vect<uint32_t>& size);
     virtual void completePlace(const uint64_t& time);
 
     virtual void update(const uint64_t& seconds) { };
-    virtual void render(Window& window, const Vect<int>& renderOffset);
+    virtual void render(Window& window, const Vect<int64_t>& renderOffset);
 
     // Save functions
     virtual std::string getSave();
     virtual std::string readSave(std::string& save);
     
     // Getter functions
-    inline bool isPlacing() const   { return placing;   }
-    inline Vect<int>& getTileSize() { return tileSize;  }
-    inline SDL_Rect& getRect()      { return renderPos; }
-    inline const ObjType& getType() { return type;      }
+    inline bool isPlacing() const        { return placing;   }
+    inline Vect<uint32_t>& getTileSize() { return tileSize;  }
+    inline SDL_Rect& getRect()           { return renderPos; }
+    inline const ObjType& getType()      { return type;      }
 
     // Setter functions
     inline void setPlacable(bool canPlace) { placable = canPlace; }
-    inline void setPos(Vect<int> pos)      { renderPos.x = pos.x; 
+    inline void setPos(Vect<int64_t> pos)      { renderPos.x = pos.x; 
                                              renderPos.y = pos.y; }
 
 protected:
-    static constexpr int alpha = 150;
+    static constexpr uint8_t alpha = 150;
 
     // SDL_Texture* image; // Add image later
-    Vect<int> tileSize; // Amount of tiles it takes up
+    Vect<uint32_t> tileSize; // Amount of tiles it takes up
     std::vector<uint8_t> renderColor;
     SDL_Rect renderPos; // Top left corner
